feat: add exception handling and apply it to post functions

+++ blog-project/src/main/java/com/gabriel/blog_project/services/PostService.java
@@ -10,6 +10,7 @@ import com.gabriel.blog_project.dtos.post.ShowPostDto;
 import com.gabriel.blog_project.dtos.post.UpdatePostDto;
 import com.gabriel.blog_project.entities.Post;
 import com.gabriel.blog_project.entities.User;
+import com.gabriel.blog_project.exceptions.EmptyDatasException;
 import com.gabriel.blog_project.repositories.PostRepository;
 import com.gabriel.blog_project.repositories.UserRepository;
 
@@ -44,7 +45,11 @@ public class PostService {
 		User user = userRepository.findById(userId)
 				.orElseThrow(() -> new RuntimeException("User not found"));
 		
-		var posts = postRepository.findAll(); 
+		var posts = postRepository.findAll();
+		
+		if(posts.isEmpty()) {
+			throw new EmptyDatasException("No posts found");
+		}
 		
 		List<ShowPostDto> result = posts.stream().map(post -> new ShowPostDto(userId, post.getTitlePost(), post.getBodyPost(), 
 				post.getImagePost(), post.getCreatedAt(), post.getUpdatedAt())).collect(Collectors.toList());
@@ -53,8 +58,9 @@ public class PostService {
 	
 	public ShowPostDto getPostById(Long id) {
 	    // Busca o post pelo id
-	    var post = postRepository.findById(id)
-	            .orElseThrow(() -> new RuntimeException("Post not found"));
+		 var post = postRepository.findById(id)
+		        .orElseThrow(() -> new EmptyDatasException("No post found with id " + id));
+	    
 
 	    // Mapeia o Post para ShowPostDto
 	    return new ShowPostDto(
@@ -74,6 +80,10 @@ public class PostService {
 		
 		var posts = postRepository.findAll();
 		
+		if (posts.isEmpty()) {
+			throw new EmptyDatasException("No posts found to delete");
+		}
+		
 		postRepository.deleteAll(posts);
 	}
 	
@@ -82,6 +92,9 @@ public class PostService {
 		User user = userRepository.findById(userId)
 				.orElseThrow(() -> new RuntimeException("User not found"));
 		
+		 var post = postRepository.findById(id)
+			        .orElseThrow(() -> new EmptyDatasException("No post found with id " + id));
+		 
 		postRepository.deleteById(id);
 	}
 	
@@ -90,13 +103,15 @@ public class PostService {
 		User user = userRepository.findById(userId)
 				.orElseThrow(() -> new RuntimeException("User not found"));
 		
-		var post = postRepository.findById(id)
-				.orElseThrow(() -> new RuntimeException("Post not found"));
+		 var post = postRepository.findById(id)
+			        .orElseThrow(() -> new EmptyDatasException("No post found with id " + id));
 		
 		post.setTitlePost(updateDto.titlePost());
 		post.setBodyPost(updateDto.bodyPost());
 		post.setImagePost(updateDto.imagePost());
 		post.setUpdatedAt(updateDto.updatedAt());
+		
+		postRepository.save(post);
 		
 		 return new ShowPostDto(
 		            post.getId(),
